add fingerprint call
