+++ connect/src/main/java/com/flip/connect/domain/repository/AccountRepository.java
@@ -1,6 +1,6 @@
 package com.flip.connect.domain.repository;
 
-import com.flip.connect.data.model.UpdateModel;
+import com.flip.connect.data.model.PatchesBase;
 import com.flip.connect.domain.model.account.AccountModel;
 import com.flip.connect.domain.boundary.CallbackBoundary;
 import com.flip.connect.domain.model.BaseResponse;
@@ -13,7 +13,7 @@ import com.flip.connect.domain.model.auth.OauthToken;
 public interface AccountRepository {
     void getAccount(OauthToken token, CallbackBoundary<AccountModel> callbackBoundary);
 
-    void updateAccount(OauthToken token, UpdateModel update, CallbackBoundary<BaseResponse> callbackBoundary);
+    void updateAccount(OauthToken token, PatchesBase update, CallbackBoundary<BaseResponse> callbackBoundary);
 
     void getAddress();
