+++ connect/src/main/java/com/flip/connect/presentation/views/login/LoginPresenter.java
@@ -11,6 +11,7 @@ import com.flip.connect.domain.model.auth.OauthToken;
 import com.flip.connect.domain.repository.LocalRepository;
 import com.flip.connect.domain.usecase.auth.OauthAccessUseCase;
 import com.flip.connect.presentation.base.BasePresenter;
+import com.flip.connect.presentation.manager.FingerPrintManager;
 
 import java.util.UUID;
 
@@ -30,7 +31,6 @@ class LoginPresenter extends BasePresenter<LoginContract.View> implements LoginC
         this.context = context;
         useCase = new OauthAccessUseCase();
         localManager = new LocalDataManager(context);
-        Log.e("token","asdfasdas");
     }
 
     @Override
@@ -40,16 +40,7 @@ class LoginPresenter extends BasePresenter<LoginContract.View> implements LoginC
             public void success(OauthToken response) {
                 Log.e("token",response.toString());
                 localManager.saveOauth(response, TokenType.ACCESS_TOKEN);
-                //region Fingerprint
-                String uniqueId = UUID.randomUUID().toString();
-                Connect.getInstance().setUniqueId(uniqueId);
-                FingerPrintLibrary.initFingerprint(context, "sandbox",
-                        "d82436c6-3664-467f-80eb-739017d13b1a", response.getUserKey(),
-                        uniqueId);
-                Log.e("UNIQUE ID", Connect.getInstance().getUniqueId());
-                FingerPrintLibrary.configFingerprint(true, true, true, true, true, true);
-                FingerPrintLibrary.getFingerprint();
-                //endregion
+                FingerPrintManager.sendFingerPrint(context, response);
                 getView().loginWithSuccess(response);
             }
